+++ src/server/routes/index.js
@@ -1,0 +1,15 @@
+#!/usr/bin/env node
+'use strict';
+const express = require('express');
+const router = express.Router();
+
+
+router.use((req,res,next) => {
+	next();
+});
+
+router.get('/', (req,res) => { 
+	res.render('index');
+});
+
+module.exports = router;
